+++ src/main/java/com/osmimport/output/actors/gdb/CompiledTableOutputActor.java
@@ -92,7 +92,7 @@ public class CompiledTableOutputActor extends MeasuredActor {
 				}
 
 				table.insertRow(r);
-				r.delete();
+				// r.delete(); 
 				
 			}