+++ ChangeLog.md
@@ -1,10 +1,62 @@
 
 #ChangeLog :
 
+- 10/10/2015 : add export command to export featureclasses and tables into geojson
 - 27/09/2015 : add multiple actions on command line
 - 06/09/2015 : version 0.6, add csv output, using base64 shape encoded geometries
 - 16/08/2015 : version 0.5,add XML/OSM file support, bugs fixes
 - 28/01/2015 : Polygon support, packaged in a simple command line, ajustment with JRE, and RAM consumpsion
+
+
+
+Version 0.7.5
+-------------
+
+- add new command : export , this permit to export in geojson fgdb tables, for small datasets, it's interesting to have a geojson, for displaying in a browser.
+
+- new commands for simplifing the script writing, for example old stream definition :
+
+<pre>
+	b = stream(osmstream, label:"streets") {
+
+		filter {
+			 e -> e instanceof OSMEntity &&
+			   e.geometryType == Geometry.Type.Polyline  && e.getFields() && e.getFields().containsKey("highway")
+		}
+		
+		transform {  e ->
+			String t = e.getFields().get("highway")
+			String name = e.getFields()?.get("name")
+			e.getFields()?.clear()
+			e.setValue("id",e.id)
+			e.setValue("name", name)
+			e.setValue("type", t)
+			return e;
+		}
+
+	}
+
+</pre>
+
+can be written with the new DSL as :
+
+<pre>
+
+	l = stream(osmstream, label:"streets") {
+		filter {
+			 e -> isPolyline(e) && has("highway") // ust be a polyline and has a highway key
+		}
+		transform { OSMEntity e ->
+			on(e).newValue("id", e.id). // define a new attribute, from an expression
+			keep("highway"). // keep the osm field highway, with the same name in the output
+			end() // generate the entity and return it
+		}
+	}
+
+
+</pre>
+
+
 
 Version 0.7
 -----------